+++ admin/settings.php
@@ -33,6 +33,21 @@ function upcart_settings_setup() {
 		'show_in_rest' => true,
 		'default'      => '$',
 	) );
+
+	register_setting( UPWPCART_PLUGIN_DOMAIN, 'upcart_redirect', array(
+		'type'         => 'string',
+		'description'  => 'Defines what page redirect after a cart interaction. Current page if empty',
+		'show_in_rest' => true,
+		'default'      => '',
+	) );
+
+	// register the POST_TYPE setting
+	register_setting( UPWPCART_PLUGIN_DOMAIN, 'upcart_auto_display', array(
+		'type'              => 'bool',
+		'description'       => 'If true, the Add to cart button will be automatically displayed in post content.',
+		'show_in_rest'      => true,
+		'default'           => false,
+	) );
 }
 
 add_action( 'admin_init', 'upcart_settings_setup' );