+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.3.0]
+### Added
+- Support to shortcodes
+- Support to form controllers
+- Added new Shortcodes
+- Added new setting: Automatic display
+### Fix
+- Bug fixes
+
+## [0.2.1]
+### Fix
+- Fix session usage (now use only when required)
+- Bug fixes
+
 
 ## [0.2.0]
 ### Added